+++ core/src/main/java/org/dcache/nfs/util/CacheMXBean.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2009 - 2012 Deutsches Elektronen-Synchroton,
+ * Copyright (c) 2009 - 2020 Deutsches Elektronen-Synchroton,
  * Member of the Helmholtz Association, (DESY), HAMBURG, GERMANY
  *
  * This library is free software; you can redistribute it and/or modify
@@ -17,20 +17,21 @@
  * details); if not, write to the Free Software Foundation, Inc.,
  * 675 Mass Ave, Cambridge, MA 02139, USA.
  */
-package org.dcache.utils;
+package org.dcache.nfs.util;
+
 
 /**
- * NOP implementation of {@link CacheEventListener}.
+ * Interface CacheMXBeanImplMXBean
  *
- * @param <T>
  */
-public class NopCacheEventListener<K, V> implements CacheEventListener<K,V> {
+public interface CacheMXBean<V> {
 
-    public void notifyPut(Cache<K,V> cache, V v) {}
-
-    public void notifyGet(Cache<K,V> cache, V v) {}
-
-    public void notifyRemove(Cache<K,V> cache, V v) {}
-
-    public void notifyExpired(Cache<K,V> cache, V v) {}
+    void clear();
+    long getEntryIdleTime();
+    long getEntryLiveTime();
+    int getSize();
+    String[] getEntries();
+    long getLastClean();
 }
+
+
