+++ core/src/main/java/org/dcache/nfs/v4/FileTracker.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 - 2019 Deutsches Elektronen-Synchroton,
+ * Copyright (c) 2017 - 2020 Deutsches Elektronen-Synchroton,
  * Member of the Helmholtz Association, (DESY), HAMBURG, GERMANY
  *
  * This library is free software; you can redistribute it and/or modify
@@ -33,7 +33,7 @@ import org.dcache.nfs.status.InvalException;
 import org.dcache.nfs.status.ShareDeniedException;
 import org.dcache.nfs.v4.xdr.stateid4;
 import org.dcache.nfs.vfs.Inode;
-import org.dcache.utils.Opaque;
+import org.dcache.nfs.util.Opaque;
 
 /**
  * A class which tracks open files.
