+++ core/src/main/java/org/dcache/nfs/v4/NFSv4StateHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2009 - 2018 Deutsches Elektronen-Synchroton,
+ * Copyright (c) 2009 - 2020 Deutsches Elektronen-Synchroton,
  * Member of the Helmholtz Association, (DESY), HAMBURG, GERMANY
  *
  * This library is free software; you can redistribute it and/or modify
@@ -44,9 +44,9 @@ import org.dcache.nfs.v4.xdr.sessionid4;
 import org.dcache.nfs.v4.xdr.stateid4;
 import org.dcache.nfs.v4.xdr.verifier4;
 import org.dcache.oncrpc4j.util.Bytes;
-import org.dcache.utils.Cache;
-import org.dcache.utils.CacheElement;
-import org.dcache.utils.NopCacheEventListener;
+import org.dcache.nfs.util.Cache;
+import org.dcache.nfs.util.CacheElement;
+import org.dcache.nfs.util.NopCacheEventListener;
 
 import static com.google.common.base.Preconditions.checkState;
